Not casting negative number to unsigned int

+++ src/main.rs
@@ -49,7 +49,7 @@ trait Sensor {
     fn sense(&self, stats_pipeline: &StatsdClient);
 }
 
-struct DummySensor(String, u64);
+struct DummySensor(String, i64);
 
 impl Sensor for DummySensor {
     fn sense(&self, statsd_client: &StatsdClient) {
@@ -57,9 +57,9 @@ impl Sensor for DummySensor {
         let place_in_interval = (now.duration_since(START_TIME).unwrap().as_secs() % PERIOD.as_secs()) as f64 / PERIOD.as_secs() as f64;
         let sin_parameter = place_in_interval * f64::consts::PI * 2.0;
         let metric_name = "test.".to_string() + &self.0;
-        let curr_value = self.1 + (self.1 as f64 * f64::sin(sin_parameter)).round() as u64;
+        let curr_value = self.1 + (self.1 as f64 * f64::sin(sin_parameter)).round() as i64;
         info!("Sense called for Sensor {}, emitting value {}", self.0, curr_value);
-        if let Err(e) = statsd_client.count(&metric_name, curr_value as i64) {
+        if let Err(e) = statsd_client.count(&metric_name, curr_value) {
             error!("Encountered and ignoring error sending stats for metric name {} and value {}: {:?}",
                    &metric_name, curr_value, e);
         }
@@ -69,7 +69,7 @@ impl Sensor for DummySensor {
 fn make_dummy_sensors(number: u8) -> Vec<DummySensor> {
     let mut dummy_sensors = Vec::new();
     for index in 0..number {
-        dummy_sensors.push(DummySensor("Sensor ".to_string() + &index.to_string(), (index as u64 + 1) * 1000));
+        dummy_sensors.push(DummySensor("Sensor ".to_string() + &index.to_string(), (index as i64 + 1) * 1000));
     }
     return dummy_sensors;
 }
